requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
streamlit-image-cropper